backend/requirements.txt
gradio
PyMuPDF
sentence-transformers
faiss-cpu
transformers
torch
